refactor: allow a product to have more than one category + move sanity folder inside src directory and update imports
