+++ package.json
@@ -21,6 +21,7 @@
     "@types/react-dom": "18.2.4",
     "antd": "^5.6.4",
     "autoprefixer": "10.4.14",
+    "dotenv": "^16.3.1",
     "next": "13.4.1",
     "next-sanity": "^4.3.2",
     "postcss": "8.4.23",
