--- sanity/config/client-config.ts
@@ -1,7 +1,0 @@
-const config = {
-  projectId: "83r3ol58",
-  dataset: "dev",
-  apiVersion: "2023-05-11",
-};
-
-export default config;
