+++ src/app/(admin)/admin/[[...index]]/page.tsx
@@ -1,8 +1,8 @@
 "use client";
 
-import { NextStudio } from 'next-sanity/studio';
+import { NextStudio } from "next-sanity/studio";
 import config from "../../../../../sanity.config";
 
 export default function AdminPage() {
-  return <NextStudio config={config} />
+  return <NextStudio config={config} />;
 }
