+++ src/components/Search.tsx
@@ -1,6 +1,6 @@
 "use client";
 import { useEffect, useState } from "react";
-import { getProducts } from "../../sanity/sanity-utils";
+import { getProducts } from "../sanity/requests/sanity-requests";
 import { QueryClient, QueryClientProvider, useQuery } from "react-query";
 import { Autocomplete, Popper } from "@mui/material";
 import SearchInput from "./SearchInput";
@@ -150,7 +150,7 @@ function Content() {
         noOptionsText={`Buscar: "${input}"`}
         options={options && input.length > 0 ? options : []}
         getOptionLabel={(o) =>
-          typeof o === "string" ? o : `${o.category} ${o.name}`
+          typeof o === "string" ? o : `${o.category.join(" ")} ${o.name}`
         }
         renderInput={(props) => (
           <SearchInput
