+++ src/sanity/config/client-config.ts
@@ -1,0 +1,7 @@
+const config = {
+  projectId: `${process.env.NEXT_PUBLIC_SANITY_PROJECT_ID}`,
+  dataset: `${process.env.NEXT_PUBLIC_SANITY_DATASET}`,
+  apiVersion: "2023-05-11",
+};
+
+export default config;
