+++ src/sanity/schemas/category-schema.ts
