+++ src/sanity/schemas/product-schema.ts
